requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.2
Pillow>=8.0.0
tqdm>=4.50.0
tensorboard>=2.4.0
opencv-python>=4.5.0
scipy>=1.6.0
open_clip_torch>=2.0.0 